syntax

+++ obliterate_engines_no_warning.sh
@@ -63,6 +63,7 @@ if test -d EnginesInstaller
 	
 	docker info >/dev/null
 	 if test $? -eq 0
+	  then
 		service docker restart
 			sleep 5
 			containers=`docker ps -a |grep -i paused |awk '{print $1}' `
@@ -78,8 +79,11 @@ if test -d EnginesInstaller
 	 
 	 	if test $keep -eq 0
 	 		then
-	 		images=`docker images -q |awk '{print $1}'`			
-	 		docker rmi $images	 		
+	 		images=`docker images -q |awk '{print $1}'`
+			if ! test -z $images
+			 then
+	 			docker rmi $images
+	 		 fi
 	 	fi
 	 	
 			service docker stop
